+++ plugins/react/frontend/components/PlaygroundList/index.js
@@ -446,7 +446,7 @@ class PlaygroundList extends Component {
               />
               <Playground
                 userFiles={this.props.userFiles}
-                dest={this.props.dest}
+                basePath={this.props.basePath}
                 commonsChunkFilename={this.props.commonsChunkFilename}
                 injectTags={this.props.injectTags}
                 component={component}
@@ -485,7 +485,7 @@ class PlaygroundList extends Component {
           variation.err ? (
             <Playground
               userFiles={this.props.userFiles}
-              dest={this.props.dest}
+              basePath={this.props.basePath}
               commonsChunkFilename={this.props.commonsChunkFilename}
               injectTags={this.props.injectTags}
               key={variationPath}
@@ -497,7 +497,7 @@ class PlaygroundList extends Component {
           ) : (
             <Playground
               userFiles={this.props.userFiles}
-              dest={this.props.dest}
+              basePath={this.props.basePath}
               commonsChunkFilename={this.props.commonsChunkFilename}
               injectTags={this.props.injectTags}
               key={variationPath}