+++ plugins/react/frontend/components/common/IFrame/index.js
@@ -4,8 +4,16 @@
 import React from 'react';
 import path from 'path';
 
-const createHtml = (componentPath, dest, userFiles, injectTags, commonsChunkFilename) => (
-  `<!DOCTYPE html>
+const createHtml = (
+  componentPath,
+  basePath = '',
+  userFiles,
+  injectTags,
+  commonsChunkFilename
+) => {
+  const iframeClientBundle = path.join(basePath, 'iframe-client-bundle.js');
+  const userBundle = path.join(basePath, 'user-bundle.js');
+  return `<!DOCTYPE html>
   <html style="height: 100%; width: 100%; margin: 0; padding: 0;">
     <head>
       ${(injectTags) ? injectTags.join('\n') : ''}
@@ -32,12 +40,12 @@ const createHtml = (componentPath, dest, userFiles, injectTags, commonsChunkFile
       <script>
         ${userFiles && userFiles.scripts.join('\n')}
       </script>
-      <script src="${(dest) ? `/${path.join(dest, 'iframe-client-bundle.js')}` : 'iframe-client-bundle.js'}"></script>
-      <script src="${(dest) ? `/${path.join(dest, 'user-bundle.js')}` : 'user-bundle.js'}"></script>
+      <script src="${iframeClientBundle}"></script>
+      <script src="${userBundle}"></script>
     </body>
   </html>
-  `
-);
+  `;
+};
 
 class IFrame extends React.Component {
 
@@ -45,7 +53,13 @@ class IFrame extends React.Component {
     const doc = this.iframe.contentDocument;
     doc.open();
     // eslint-disable-next-line max-len
-    doc.write(createHtml(this.props.componentPath, this.props.dest, this.props.userFiles, this.props.injectTags, this.props.commonsChunkFilename));
+    doc.write(createHtml(
+      this.props.componentPath,
+      this.props.basePath,
+      this.props.userFiles,
+      this.props.injectTags,
+      this.props.commonsChunkFilename
+    ));
     doc.close();
 
     this.iframe.contentWindow.INITIAL_COMPONENT_DATA = this.props.variationProps;