+++ plugins/react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 CarteBlanche is maintained by [@mxstbr](https://github.com/mxstbr), [@nikgraf](https://github.com/nikgraf), [@jantimon](https://github.com/jantimon) and an amazing team of contributors. All contributions, except for those by Max, Nik and Timon, are credited here.
 
+## to be released
+
+- changed; switch to basePath support instead of dest
+
 ## v0.2 / 2016-06-12
 
 - added; show JSX code of variation in interface, thanks [@RStankov](https://github.com/RStankov)